refactor: 숫자, 문자 형식 재조정 및 예외처리 클래스 세분화
