+++ src/main/java/vendingmachine/controller/NumberController.java
@@ -2,7 +2,7 @@ package vendingmachine.controller;
 
 import vendingmachine.util.Constant;
 
-public class ExceptionController {
+public class NumberController {
 
 	public static void isInteger(String input) {
 		try {
@@ -28,18 +28,6 @@ public class ExceptionController {
 	public static void isMultipleOfTen(int input) {
 		if (input % 10 != 0) {
 			throw new IllegalArgumentException(Constant.ERROR_IS_NOT_MULTIPLE_OF_TEN);
-		}
-	}
-
-	public static void isInfoFormatValidate(String input) {
-		if (input.charAt(0) != '[' || input.charAt(input.length() - 1) != ']') {
-			throw new IllegalArgumentException(Constant.ERROR_IS_NOT_VALIDATE_INFO_FORMAT);
-		}
-	}
-
-	public static void isNumberOfInfo3(String[] input) {
-		if (input.length != 3) {
-			throw new IllegalArgumentException(Constant.ERROR_IS_NOT_3_INFO);
 		}
 	}
 }
